Move some functions/vars into different files. Shallow refactor
