+++ gameinfo.lua
@@ -1,0 +1,1 @@
+PlayTime = 0
