+++ settings.lua
@@ -1,0 +1,8 @@
+ShowProgress = true
+
+CellAmount = 10
+
+---@type Direction
+Rotation = Direction.RIGHT
+---@type CellType
+BuildSelection = CellType.NONE
